requirements.txt
streamlit
pandas
numpy
opencv-python-headless
Pillow
requests